focuslog/requirements-gui.txt
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0
pydantic>=2.8,<3.0
pywebview>=5.1,<6.0
